requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==22.0.0
python-dotenv==1.0.1
requests==2.32.3
PyJWT==2.8.0
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
google-generativeai==0.7.1
google-api-core==2.19.1

# Sub-dependencies (auto-generated by pip freeze)
appdirs==1.4.4
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
frozendict==2.4.4
google-ai-generativelanguage==0.6.6
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
multitasking==0.0.11
peewee==3.17.5
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
rsa==4.9
soupsieve==2.5
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
